refactor: `Proxy`에서 직접 `DOM`을 조작하지 않도록 수정
