+++ src/js/constants/constraints.js
@@ -1,0 +1,1 @@
+export const MAX_NUMBER_LENGTH = 3;
