+++ src/js/constants/events.js
@@ -1,0 +1,1 @@
+export const CHANGE_DISPLAY_VALUE = 'CHANGE_DISPLAY_VALUE';
