+++ src/js/constants/inputs.js
@@ -1,0 +1,11 @@
+export const OPERATORS = {
+	ADDITION: '+',
+	SUBTRACTION: '-',
+	MULTIPLICATION: 'X',
+	DIVISION: '/',
+	EQUALS: '=',
+};
+
+export const MODIFIERS = {
+	ALL_CLEAR: 'AC',
+};
