+++ src/js/constants/messages.js
@@ -1,0 +1,2 @@
+export const ALERT_MAX_NUMBER_LENGTH_MESSAGE =
+	'숫자는 세 자리까지만 입력 가능합니다!';
